Migrated from Java 5  (#3464)
* Replace with enhanced for

* Replace indexOf with contains

* StringBulder for String Buffer

* StringBuilder for StringBuffer

* Unboxing

* More Unboxing

* replace while loop with enhanced for

* Spotless

+++ src/net/sourceforge/kolmafia/session/PvpManager.java
@@ -92,7 +92,7 @@ public class PvpManager {
 
     String name = matchingNames.get(0);
     Integer stance = PvpManager.canonicalStanceToOption.get(name);
-    return stance == null ? -1 : stance.intValue();
+    return stance == null ? -1 : stance;
   }
 
   public static final String findStance(final int stance) {
